% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFun.R
\name{ageFun}
\alias{ageFun}
\title{ageFun}
\usage{
ageFun(born, date = Sys.time(), format = "\%Y-\%m-\%d",
  include.months = TRUE)
}
\arguments{
\item{born}{is a character or date class value/vector representing the date of birth}

\item{date}{is a character or date class value/vector representing the date to calculate age}

\item{format}{is a character string the format for character to date conversion (as in as.Date() function)}

\item{include.months}{is a logical to decide if months are shown in results}
}
\value{
a list including age in years and months (if specified)
}
\description{
Calculates age in years
}
\examples{
ageFun(born = "1900-02-01",date = "1950-01-01",format = "\%Y-\%m-\%d",include.months = FALSE)

}
\author{
Pablo O Rodriguez
}
