% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builtTable.R
\name{buildTable}
\alias{buildTable}
\title{Summary table with statistical tests for groups comparison}
\usage{
buildTable(data, group = 1, non.parametric = FALSE)
}
\arguments{
\item{data}{is a vector with data}

\item{group}{is a factor vector for grouping}

\item{non.parametric}{is a logical vector. FALSE allows the function to choose the test according to the distribution. If TRUE, the function uses non parametric tests by default.}
}
\description{
Creates a summary statistics table of data comparing by group with the appropiated statistical test
}
\examples{
buildTable (mtcars$mpg, mtcars$cyl, non.parametric=FALSE)
}
