% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFun.R
\name{is.normal}
\alias{is.normal}
\title{is.normal}
\usage{
is.normal(data, group = NA)
}
\arguments{
\item{data}{is a numeric vector}

\item{group}{is a vector with grouping data}
}
\value{
logical value, TRUE if normality is verified in all groups according to shapiro test P
}
\description{
Calculates shapiro test in data by groups
}
\examples{
is.normal (mtcars$mpg, mtcars$cyl)

}
\author{
Pablo O Rodriguez
}
