% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFun.R
\name{meanSD}
\alias{meanSD}
\title{meanSD}
\usage{
meanSD(data)
}
\arguments{
\item{data}{is a numeric vector}
}
\value{
the mean plus minus SD of data
}
\description{
Calculates mean plus minus SD from a numeric vector
}
\examples{
meanSD (mtcars$mpg)

}
\author{
Pablo O Rodriguez
}
