% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFun.R
\name{medianIQR}
\alias{medianIQR}
\title{medianIQR}
\usage{
medianIQR(data)
}
\arguments{
\item{data}{is a numeric vector}
}
\value{
the median (IQR) of data
}
\description{
Calculates mean and interquartile range from a numeric vector
}
\examples{
medianIQR (mtcars$cyl)

}
\author{
Pablo O Rodriguez
}
