% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pCalc.R
\name{p.Anova}
\alias{p.Anova}
\title{One way ANOVA P of data x group}
\usage{
p.Anova(data, group)
}
\arguments{
\item{data}{is a numeric or integer vector with data}

\item{group}{is a factor vector for grouping}
}
\description{
Calculates one way ANOVA P (two tails) of data by group
}
\examples{
p.Anova (mtcars$mpg, mtcars$cyl)
}
