% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pCalc.R
\name{p.Chisq}
\alias{p.Chisq}
\title{Chi square P of data x group}
\usage{
p.Chisq(data, group)
}
\arguments{
\item{data}{is a factor vector with data}

\item{group}{is a factor vector for grouping}
}
\description{
Calculates Chi square test P (two tails) of data by group
}
\examples{
p.Chisq (mtcars$mpg, mtcars$cyl)
}
