% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pCalc.R
\name{p.Fisher}
\alias{p.Fisher}
\title{Fisher exact P of data x group}
\usage{
p.Fisher(data, group)
}
\arguments{
\item{data}{is a factor vector with data}

\item{group}{is a factor vector for grouping}
}
\description{
Calculates Fisher exact test P (two tails) of data by group
}
\examples{
p.Fisher (mtcars$mpg, mtcars$cyl)
}
