% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pCalc.R
\name{p.Kruskal}
\alias{p.Kruskal}
\title{Kruskal-Wallis rank sum test P of data x group}
\usage{
p.Kruskal(data, group)
}
\arguments{
\item{data}{is a numeric or integer vector with data}

\item{group}{is a factor vector for grouping}
}
\description{
Calculates Kruskal Wallis test P (two tails) of data by group
}
\examples{
p.Kruskal (mtcars$mpg, mtcars$cyl)
}
