% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pCalc.R
\name{p.Ttest}
\alias{p.Ttest}
\title{T test P of data x group}
\usage{
p.Ttest(data, group)
}
\arguments{
\item{data}{is a numeric or integer vector with data}

\item{group}{is a factor vector for grouping}
}
\description{
Calculates t test P (two tails) of data by group
}
\examples{
p.Ttest (mtcars$mpg, mtcars$cyl)
}
