% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicFun.R
\name{shapiroGroup}
\alias{shapiroGroup}
\title{shapiroGroup}
\usage{
shapiroGroup(data, group = NA)
}
\arguments{
\item{data}{is a numeric vector}

\item{group}{is a vector with grouping data}
}
\value{
the shapiro test output of data for each group
}
\description{
Calculates shapiro test in data by groups
}
\examples{
shapiroGroup (mtcars$mpg, mtcars$cyl)

}
\author{
Pablo O Rodriguez
}
